[BREAKING CHANGE] Support 8-Bit pcm in all sound sources
This brings support for 8-Bit pcm data in SoundBuffer and
StreamedSoundSource. The public SoundBuffer constructor had to be
changed as well as the AudioStream interface, hence the [BREAKING
CHANGE].

Thanks @Frosty-J for the groundwork in libGDX PR 6792, which left me
baffled as to why I braindeadly ignored this unnecessary restriction and
which ultimately led to this commit.

The support of other formats like in 6792 is not subject of THIS commit.

Signed-off-by: Matthias <webmaster@pottgames.de>